Add a intermediate create for shared types with WASM.
Had to add an additional layer or proxies, so the shared types did not
have to reference, the AWS sdks, which don't compile for WASM due to the
tokio dependency.

Every shared types gets a Proxy type implementing the From trait to
convert from the AWS -> Proxy type and should implement the Deref trait
to make accessing the shared type easier.